requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r ./base.txt

gunicorn==20.0.4  # https://github.com/benoitc/gunicorn
#psycopg2-binary==2.8.5  # https://github.com/psycopg/psycopg2
sentry-sdk==0.14.3  # https://github.com/getsentry/sentry-python

# Django
# ------------------------------------------------------------------------------
django-anymail[mailgun]==7.0.0  # https://github.com/anymail/django-anymail
django-cloudinary-storage
